# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST Panel de Control de Club Deportivo",
        "description": "API para autenticación y gestión de un club deportivo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/v1/reservations": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener lista de reservas",
                "description": "Devuelve un listado paginado de reservas con información del socio y la pista",
                "operationId": "7fce57eed139dca21be9f8f3bce67d70",
                "responses": {
                    "200": {
                        "description": "Lista de reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReservationResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Crear una nueva reserva",
                "description": "Registra una reserva en la base de datos",
                "operationId": "fc082e16ea2f3a97f378e946ac6b7b46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener una reserva por ID",
                "description": "Devuelve los datos de una reserva específica junto con información del socio y la pista",
                "operationId": "695883ff60d9eedd252eadd9cf3f53cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Actualizar una reserva",
                "description": "Modifica una reserva existente",
                "operationId": "d756f9cb19a6fe051557d6e84019125a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Eliminar una reserva",
                "description": "Borra una reserva usando su ID",
                "operationId": "ec63c6b26bd88738be38deb8fca0c6eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Reserva eliminada correctamente"
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Actualizar parcialmente una reserva",
                "description": "Modifica solo los campos enviados en la solicitud",
                "operationId": "3aa4b8ba37f14d9a21862d887543dde2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reservations/by-day": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener reservas por día",
                "description": "Devuelve todas las reservas de una fecha específica",
                "operationId": "a82c6552c08f352498c9712064601f59",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha en formato YYYY-MM-DD",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-06-15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de reservas en la fecha especificada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reservations": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReservationResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de la fecha"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreReservationRequest": {
                "title": "Store Reservation Request",
                "description": "Datos necesarios para crear una reserva",
                "required": [
                    "member_id",
                    "court_id",
                    "date",
                    "time"
                ],
                "properties": {
                    "member_id": {
                        "description": "ID del miembro",
                        "type": "integer",
                        "example": 1
                    },
                    "court_id": {
                        "description": "ID de la pista",
                        "type": "integer",
                        "example": 3
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-15"
                    },
                    "time": {
                        "type": "string",
                        "example": "14:00"
                    }
                },
                "type": "object"
            },
...
            "UpdateReservationRequest": {
                "title": "Update Reservation Request",
                "description": "Datos para actualizar una reserva",
                "properties": {
                    "member_id": {
                        "description": "ID del miembro",
                        "type": "integer",
                        "example": 1
                    },
                    "court_id": {
                        "description": "ID de la pista",
                        "type": "integer",
                        "example": 3
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-15"
                    },
                    "time": {
                        "type": "string",
                        "example": "14:00"
                    }
                },
                "type": "object"
            },
...
            "ReservationResource": {
                "title": "Reservation Resource",
                "description": "Representación de una reserva en la API",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "member": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 10
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            }
                        },
                        "type": "object"
                    },
                    "court": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 5
                            },
                            "name": {
                                "type": "string",
                                "example": "Court A"
                            }
                        },
                        "type": "object"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-15"
                    },
                    "time": {
                        "type": "string",
                        "example": "14:00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-14T12:05:09Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-14T12:10:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}